Add map functionality (Exercise 2.6).
It would be neat to try and amalgamate the functionality of sets and maps
but for now I'm going to leave them separate.

The optimisations to not create copies of things for non-mutating changes
and to use d+1 rather than 2d comparisons for "insert" have not been done,
largely because they'd be complex.